requirements.txt
flask
twilio
google-generativeai
gspread
google-auth
python-dotenv
